Updated Nav and Home Page What have I done lol
