+++ client/src/components/ContactUs.js
@@ -2,7 +2,7 @@ import "./stylesheets/contactUs.css"
 
 export const ContactUs = () => {
     return (
-        <div>
+        <div className="contactus_div">
             <h1>Contact Us Page</h1>
             <img src="https://i.ibb.co/60SCrH3/Website-Under-Construction.gif" className="funny_gif" alt="Website Under Construction GIF" />
         </div>
