+++ client/src/components/History.js
@@ -2,7 +2,7 @@ import "./stylesheets/history.css"
 
 export const History = () => {
     return (
-        <div>
+        <div className="history_div">
             <h1>History Page</h1>
             <img src="https://i.ibb.co/60SCrH3/Website-Under-Construction.gif" className="funny_gif" alt="Website Under Construction GIF" />
         </div>
