--- client/src/components/NavBarButton.js
@@ -1,24 +1,0 @@
-import { Link } from "react-router-dom"
-import "./stylesheets/navBarButton.css"
-
-const STYLES = ["btn--primary", "btn--outline"]
-
-const SIZES = ["btn--medium", "btn--large"]
-
-export const NavBarButton = ({ children, type, onClick, buttonStyle, buttonSize }) => {
-    const checkButtonStyle = STYLES.includes(buttonStyle) ? buttonStyle : STYLES[0]
-
-    const checkButtonSize = SIZES.includes(buttonSize) ? buttonSize : SIZES[0]
-
-    return (
-        <Link to="/login" className="btn-mobile">
-            <button
-                className={`btn ${checkButtonStyle} ${checkButtonSize}`}
-                onClick={onClick}
-                type={type}
-            >
-                {children}
-            </button>
-        </Link>
-    )
-}
