+++ client/src/components/stylesheets/contactUs.css
@@ -1,4 +1,4 @@
-div {
+.contactus_div {
     text-align: center;
 }
