+++ client/src/components/stylesheets/history.css
@@ -1,4 +1,4 @@
-div {
+.history_div {
     text-align: center;
 }
